# HW8/src/api/generated.go
// Package api provides primitives to interact with the openapi HTTP API.
//
// Code generated by github.com/oapi-codegen/oapi-codegen/v2 version v2.5.0 DO NOT EDIT.
package api

import (
	"fmt"
	"net/http"
	"time"

	"github.com/labstack/echo/v4"
	"github.com/oapi-codegen/runtime"
)

const (
	ApiKeyAuthScopes = "ApiKeyAuth.Scopes"
	BearerAuthScopes = "BearerAuth.Scopes"
)

// CartItem defines model for CartItem.
type CartItem struct {
	Product *Product `json:"product,omitempty"`

	// ProductId Product identifier
	ProductId int32 `json:"product_id"`

	// Quantity Quantity of this product in cart
	Quantity int32 `json:"quantity"`
}

// Error defines model for Error.
type Error struct {
	// Details Additional error details
	Details *string `json:"details,omitempty"`

	// Error Error code
	Error string `json:"error"`

	// Message Human-readable error message
	Message string `json:"message"`
}

// Product defines model for Product.
type Product struct {
	// CategoryId Product category identifier
	CategoryId int32 `json:"category_id"`

	// Manufacturer Product manufacturer name
	Manufacturer string `json:"manufacturer"`

	// ProductId Unique identifier for the product
	ProductId int32 `json:"product_id"`

	// Sku Stock Keeping Unit - unique product code
	Sku string `json:"sku"`

	// SomeOtherId Additional identifier for product
	SomeOtherId int32 `json:"some_other_id"`

	// Weight Product weight in grams
	Weight int32 `json:"weight"`
}

// ShoppingCart defines model for ShoppingCart.
type ShoppingCart struct {
	// CreatedAt When cart was created
	CreatedAt *time.Time `json:"created_at,omitempty"`

	// CustomerId Customer who owns this cart
	CustomerId int32 `json:"customer_id"`

	// Items Items in the shopping cart
	Items []CartItem `json:"items"`

	// ShoppingCartId Unique identifier for the shopping cart
	ShoppingCartId int32 `json:"shopping_cart_id"`
}

// ProcessPaymentJSONBody defines parameters for ProcessPayment.
type ProcessPaymentJSONBody struct {
	// CreditCardNumber Credit card number (13-19 digits)
	CreditCardNumber string `json:"credit_card_number"`

	// ShoppingCartId Unique identifier for the shopping cart
	ShoppingCartId int32 `json:"shopping_cart_id"`
}

// CreateShoppingCartJSONBody defines parameters for CreateShoppingCart.
type CreateShoppingCartJSONBody struct {
	// CustomerId Unique identifier for the customer
	CustomerId int32 `json:"customer_id"`
}

// AddItemsToCartJSONBody defines parameters for AddItemsToCart.
type AddItemsToCartJSONBody struct {
	// ProductId Unique identifier for the product
	ProductId int32 `json:"product_id"`

	// Quantity Number of items to add
	Quantity int32 `json:"quantity"`
}

// ReserveInventoryJSONBody defines parameters for ReserveInventory.
type ReserveInventoryJSONBody struct {
	// ProductId Unique identifier for the product
	ProductId int32 `json:"product_id"`

	// Quantity Quantity to reserve
	Quantity int32 `json:"quantity"`
}

// ShipProductJSONBody defines parameters for ShipProduct.
type ShipProductJSONBody struct {
	// ProductId Unique identifier for the product
	ProductId int32 `json:"product_id"`

	// Quantity Quantity to ship
	Quantity int32 `json:"quantity"`
}

// ProcessPaymentJSONRequestBody defines body for ProcessPayment for application/json ContentType.
type ProcessPaymentJSONRequestBody ProcessPaymentJSONBody

// AddProductDetailsJSONRequestBody defines body for AddProductDetails for application/json ContentType.
type AddProductDetailsJSONRequestBody = Product

// CreateShoppingCartJSONRequestBody defines body for CreateShoppingCart for application/json ContentType.
type CreateShoppingCartJSONRequestBody CreateShoppingCartJSONBody

// AddItemsToCartJSONRequestBody defines body for AddItemsToCart for application/json ContentType.
type AddItemsToCartJSONRequestBody AddItemsToCartJSONBody

// ReserveInventoryJSONRequestBody defines body for ReserveInventory for application/json ContentType.
type ReserveInventoryJSONRequestBody ReserveInventoryJSONBody

// ShipProductJSONRequestBody defines body for ShipProduct for application/json ContentType.
type ShipProductJSONRequestBody ShipProductJSONBody

// ServerInterface represents all server handlers.
type ServerInterface interface {
	// Process credit card payment
	// (POST /payments/checkout)
	ProcessPayment(ctx echo.Context) error
	// Get product by ID
	// (GET /products/{productId})
	GetProduct(ctx echo.Context, productId int32) error
	// Add product details
	// (POST /products/{productId}/details)
	AddProductDetails(ctx echo.Context, productId int32) error
	// Create a new shopping cart
	// (POST /shopping-carts)
	CreateShoppingCart(ctx echo.Context) error
	// Get shopping cart
	// (GET /shopping-carts/{shoppingCartId})
	GetShoppingCart(ctx echo.Context, shoppingCartId int32) error
	// Checkout shopping cart
	// (POST /shopping-carts/{shoppingCartId}/checkout)
	CheckoutCart(ctx echo.Context, shoppingCartId int32) error
	// Add items to shopping cart
	// (POST /shopping-carts/{shoppingCartId}/items)
	AddItemsToCart(ctx echo.Context, shoppingCartId int32) error
	// Reserve product inventory
	// (POST /warehouse/reserve)
	ReserveInventory(ctx echo.Context) error
	// Ship product
	// (POST /warehouse/ship)
	ShipProduct(ctx echo.Context) error
}

// ServerInterfaceWrapper converts echo contexts to parameters.
type ServerInterfaceWrapper struct {
	Handler ServerInterface
}

// ProcessPayment converts echo context to params.
func (w *ServerInterfaceWrapper) ProcessPayment(ctx echo.Context) error {
	var err error

	ctx.Set(ApiKeyAuthScopes, []string{})

	ctx.Set(BearerAuthScopes, []string{})

	// Invoke the callback with all the unmarshaled arguments
	err = w.Handler.ProcessPayment(ctx)
	return err
}

// GetProduct converts echo context to params.
func (w *ServerInterfaceWrapper) GetProduct(ctx echo.Context) error {
	var err error
	// ------------- Path parameter "productId" -------------
	var productId int32

	err = runtime.BindStyledParameterWithOptions("simple", "productId", ctx.Param("productId"), &productId, runtime.BindStyledParameterOptions{ParamLocation: runtime.ParamLocationPath, Explode: false, Required: true})
	if err != nil {
		return echo.NewHTTPError(http.StatusBadRequest, fmt.Sprintf("Invalid format for parameter productId: %s", err))
	}

	ctx.Set(ApiKeyAuthScopes, []string{})

	ctx.Set(BearerAuthScopes, []string{})

	// Invoke the callback with all the unmarshaled arguments
	err = w.Handler.GetProduct(ctx, productId)
	return err
}

// AddProductDetails converts echo context to params.
func (w *ServerInterfaceWrapper) AddProductDetails(ctx echo.Context) error {
	var err error
	// ------------- Path parameter "productId" -------------
	var productId int32

	err = runtime.BindStyledParameterWithOptions("simple", "productId", ctx.Param("productId"), &productId, runtime.BindStyledParameterOptions{ParamLocation: runtime.ParamLocationPath, Explode: false, Required: true})
	if err != nil {
		return echo.NewHTTPError(http.StatusBadRequest, fmt.Sprintf("Invalid format for parameter productId: %s", err))
	}

	ctx.Set(ApiKeyAuthScopes, []string{})

	ctx.Set(BearerAuthScopes, []string{})

	// Invoke the callback with all the unmarshaled arguments
	err = w.Handler.AddProductDetails(ctx, productId)
	return err
}

// CreateShoppingCart converts echo context to params.
func (w *ServerInterfaceWrapper) CreateShoppingCart(ctx echo.Context) error {
	var err error

	ctx.Set(ApiKeyAuthScopes, []string{})

	ctx.Set(BearerAuthScopes, []string{})

	// Invoke the callback with all the unmarshaled arguments
	err = w.Handler.CreateShoppingCart(ctx)
	return err
}

// GetShoppingCart converts echo context to params.
func (w *ServerInterfaceWrapper) GetShoppingCart(ctx echo.Context) error {
	var err error
	// ------------- Path parameter "shoppingCartId" -------------
	var shoppingCartId int32

	err = runtime.BindStyledParameterWithOptions("simple", "shoppingCartId", ctx.Param("shoppingCartId"), &shoppingCartId, runtime.BindStyledParameterOptions{ParamLocation: runtime.ParamLocationPath, Explode: false, Required: true})
	if err != nil {
		return echo.NewHTTPError(http.StatusBadRequest, fmt.Sprintf("Invalid format for parameter shoppingCartId: %s", err))
	}

	ctx.Set(ApiKeyAuthScopes, []string{})

	ctx.Set(BearerAuthScopes, []string{})

	// Invoke the callback with all the unmarshaled arguments
	err = w.Handler.GetShoppingCart(ctx, shoppingCartId)
	return err
}

// CheckoutCart converts echo context to params.
func (w *ServerInterfaceWrapper) CheckoutCart(ctx echo.Context) error {
	var err error
	// ------------- Path parameter "shoppingCartId" -------------
	var shoppingCartId int32

	err = runtime.BindStyledParameterWithOptions("simple", "shoppingCartId", ctx.Param("shoppingCartId"), &shoppingCartId, runtime.BindStyledParameterOptions{ParamLocation: runtime.ParamLocationPath, Explode: false, Required: true})
	if err != nil {
		return echo.NewHTTPError(http.StatusBadRequest, fmt.Sprintf("Invalid format for parameter shoppingCartId: %s", err))
	}

	ctx.Set(ApiKeyAuthScopes, []string{})

	ctx.Set(BearerAuthScopes, []string{})

	// Invoke the callback with all the unmarshaled arguments
	err = w.Handler.CheckoutCart(ctx, shoppingCartId)
	return err
}

// AddItemsToCart converts echo context to params.
func (w *ServerInterfaceWrapper) AddItemsToCart(ctx echo.Context) error {
	var err error
	// ------------- Path parameter "shoppingCartId" -------------
	var shoppingCartId int32

	err = runtime.BindStyledParameterWithOptions("simple", "shoppingCartId", ctx.Param("shoppingCartId"), &shoppingCartId, runtime.BindStyledParameterOptions{ParamLocation: runtime.ParamLocationPath, Explode: false, Required: true})
	if err != nil {
		return echo.NewHTTPError(http.StatusBadRequest, fmt.Sprintf("Invalid format for parameter shoppingCartId: %s", err))
	}

	ctx.Set(ApiKeyAuthScopes, []string{})

	ctx.Set(BearerAuthScopes, []string{})

	// Invoke the callback with all the unmarshaled arguments
	err = w.Handler.AddItemsToCart(ctx, shoppingCartId)
	return err
}

// ReserveInventory converts echo context to params.
func (w *ServerInterfaceWrapper) ReserveInventory(ctx echo.Context) error {
	var err error

	ctx.Set(ApiKeyAuthScopes, []string{})

	ctx.Set(BearerAuthScopes, []string{})

	// Invoke the callback with all the unmarshaled arguments
	err = w.Handler.ReserveInventory(ctx)
	return err
}

// ShipProduct converts echo context to params.
func (w *ServerInterfaceWrapper) ShipProduct(ctx echo.Context) error {
	var err error

	ctx.Set(ApiKeyAuthScopes, []string{})

	ctx.Set(BearerAuthScopes, []string{})

	// Invoke the callback with all the unmarshaled arguments
	err = w.Handler.ShipProduct(ctx)
	return err
}

// This is a simple interface which specifies echo.Route addition functions which
// are present on both echo.Echo and echo.Group, since we want to allow using
// either of them for path registration
type EchoRouter interface {
	CONNECT(path string, h echo.HandlerFunc, m ...echo.MiddlewareFunc) *echo.Route
	DELETE(path string, h echo.HandlerFunc, m ...echo.MiddlewareFunc) *echo.Route
	GET(path string, h echo.HandlerFunc, m ...echo.MiddlewareFunc) *echo.Route
	HEAD(path string, h echo.HandlerFunc, m ...echo.MiddlewareFunc) *echo.Route
	OPTIONS(path string, h echo.HandlerFunc, m ...echo.MiddlewareFunc) *echo.Route
	PATCH(path string, h echo.HandlerFunc, m ...echo.MiddlewareFunc) *echo.Route
	POST(path string, h echo.HandlerFunc, m ...echo.MiddlewareFunc) *echo.Route
	PUT(path string, h echo.HandlerFunc, m ...echo.MiddlewareFunc) *echo.Route
	TRACE(path string, h echo.HandlerFunc, m ...echo.MiddlewareFunc) *echo.Route
}

// RegisterHandlers adds each server route to the EchoRouter.
func RegisterHandlers(router EchoRouter, si ServerInterface) {
	RegisterHandlersWithBaseURL(router, si, "")
}

// Registers handlers, and prepends BaseURL to the paths, so that the paths
// can be served under a prefix.
func RegisterHandlersWithBaseURL(router EchoRouter, si ServerInterface, baseURL string) {

	wrapper := ServerInterfaceWrapper{
		Handler: si,
	}

	router.POST(baseURL+"/payments/checkout", wrapper.ProcessPayment)
	router.GET(baseURL+"/products/:productId", wrapper.GetProduct)
	router.POST(baseURL+"/products/:productId/details", wrapper.AddProductDetails)
	router.POST(baseURL+"/shopping-carts", wrapper.CreateShoppingCart)
	router.GET(baseURL+"/shopping-carts/:shoppingCartId", wrapper.GetShoppingCart)
	router.POST(baseURL+"/shopping-carts/:shoppingCartId/checkout", wrapper.CheckoutCart)
	router.POST(baseURL+"/shopping-carts/:shoppingCartId/items", wrapper.AddItemsToCart)
	router.POST(baseURL+"/warehouse/reserve", wrapper.ReserveInventory)
	router.POST(baseURL+"/warehouse/ship", wrapper.ShipProduct)

}